fix: client image loading fine in client details page and in list
